src/requirements.txt
pyalsaaudio
pyyaml
numpy
sounddevice
opencv-python